#217 refactor:sendMessage customHook으로 관리
