+++ src/components/chat/chatDetail/ChatSend.tsx
@@ -2,18 +2,26 @@ import { useState } from 'react';
 
 import Icon from '@/components/common/Icon';
 import { Input } from '@/components/common/Input/Input';
+import { useChatProps } from '@/hooks/useChat';
+import { useSendMessage } from '@/hooks/useSendMessage';
 import * as S from '../Chat.styles';
 
-type sendMessageProps = {
-  sendMessage: (e: string) => void;
-};
-
-const ChatSend = ({ sendMessage }: sendMessageProps) => {
+const ChatSend = ({ customRoomId, stompClient, role, user, product, seller }: useChatProps) => {
   const [text, setText] = useState<string>('');
 
   const inputHandle = (e: string) => {
     setText(e);
   };
+
+  const sendMessage = useSendMessage({
+    customRoomId,
+    // text,
+    stompClient,
+    role,
+    user,
+    product,
+    seller,
+  });
 
   const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault(); // 폼 제출 시 페이지 새로고침 방지
